+++ workers/tests/integrationtests/test_indicator_last_edit.py
@@ -13,13 +13,13 @@ from .utils import oqt_vcr
 class TestIndicatorLastEdit(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test(self):
-        infile = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "fixtures",
-            "heidelberg_altstadt.geojson",
+        # Heidelberg
+        bpolys = asyncio.run(
+            db_client.get_bpolys_from_db(
+                dataset="regions", feature_id=3, fid_field="ogc_fid"
+            )
         )
-        with open(infile, "r") as f:
-            bpolys = geojson.load(f)
+
         indicator = LastEdit(bpolys=bpolys, layer_name="major_roads_count")
         asyncio.run(indicator.preprocess())
         indicator.calculate()
@@ -42,18 +42,22 @@ class TestIndicatorLastEdit(unittest.TestCase):
     #     self.assertLess(indicator.total_features, indicator.edited_features)
     #     self.assertEqual(indicator.share_edited_features, 100)
 
+    # TODO: Add GeoJSON fixture with regions covering the requirements of this test
     @oqt_vcr.use_cassette()
     def test_no_amenities(self):
         """Test area with no amenities"""
-        dataset = "regions"
-        feature_id = 28  # Niger Kanan Bakache
-        bpolys = asyncio.run(
-            db_client.get_bpolys_from_db(dataset, feature_id, "ogc_fid")
+        infile = os.path.join(
+            os.path.dirname(os.path.abspath(__file__)),
+            "fixtures",
+            "niger-kanan-bakache.geojson",
         )
+        with open(infile, "r") as f:
+            bpolys = geojson.load(f)
 
         indicator = LastEdit(layer_name="amenities", bpolys=bpolys)
         asyncio.run(indicator.preprocess())
         self.assertEqual(indicator.total_features, 0)
+
         indicator.calculate()
         self.assertEqual(indicator.result.label, "undefined")
         self.assertEqual(indicator.result.value, None)